% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_clean_combined_data.R
\name{get_data}
\alias{get_data}
\title{high school academia data}
\usage{
get_data(path)
}
\arguments{
\item{path}{character : directory of the files}
}
\value{
data.tables containing the contents of csv files
}
\description{
high school academia data
}
\examples{
data_dt <- get_data("~/data.csv")
}
