% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_clean_combined_data.R
\name{load_high_school_data}
\alias{load_high_school_data}
\title{clean and load high school education data}
\usage{
load_high_school_data(grant_data_path, senior_marks_data_path, sample_n = 3000)
}
\arguments{
\item{grant_data_path}{Character. Path to the grant data CSV.}

\item{sample_n}{Integer. Number of rows to sample for preview (optional).}

\item{Character.}{Path to the senior marks data CSV.}
}
\value{
A sampled subset of the data
}
\description{
clean and load high school education data
}
\examples{
#' result <- load_high_school_data("grant.csv", "marks.csv", 3000)
}
